further BORLAND compiler support; Blended model minor QA/QC fix [BENCHMARKED]
BORLAND C compiler support (for Pierre @CMM)
-changed GetCurrentTime() to GetCurrentMachineTime() (RavenInclude.h, CommonFunctions.cpp; ConstituentModel.cpp; StandardOutput.cpp; CustomOutput.cpp)
-deleted remaining references to reserved words OUT, IN, FILENAME (ConstituentModel.cpp;ParseGWFile.cpp;UnitTesting.cpp)
-minor fix to ModelForcingGrids.cpp for type casting

QA/QC
-added warnings to BlendedWeights commands (ParseInput.cpp)

git-svn-id: https://chyms.nrc.gc.ca/svn/raven/trunk@344 9a3899ca-7adb-4d27-9b38-7f4ebb3123b0

+++ src/ModelForcingGrids.cpp
@@ -337,7 +337,7 @@ void CModel::GenerateMinMaxSubdailyTempFromAve(const optStruct &Options)
   pTave_daily->Initialize(Options);//  needed for correct mapping from time series to model time
 
   double interval = pTave_daily->GetInterval();
-  int       nVals = (int)ceil(pTave_daily->GetChunkSize()); // number of subdaily values (input resolution) - should be 1
+  int       nVals = (int)ceil((float)(pTave_daily->GetChunkSize())); // number of subdaily values (input resolution) - should be 1
   
   pTmin_daily = ForcingCopyCreate(pTave_daily,F_TEMP_DAILY_MIN,interval,nVals);
   pTmax_daily = ForcingCopyCreate(pTave_daily,F_TEMP_DAILY_MAX,interval,nVals);
